Test navigating from table to experiment page
